segregated header file definitions into cpp files

+++ radix/src/game.cpp
@@ -1,16 +1,15 @@
 #include <iostream>
 
-#include "./map.h"
-#include "./game.h"
-#include "./asset_manager.h"
-#include "./entity_manager.h"
-#include "../lib/glm/glm.hpp"
-#include "./sprite_component.h"
-#include "./keyboard_component.h"
-#include "./collider_component.h"
-#include "./transform_component.h"
-#include "./text_label_component.h"
-#include "./projectile_emitter_component.h"
+#include "map.h"
+#include "game.h"
+#include "asset_manager.h"
+#include "entity_manager.h"
+#include "sprite_component.h"
+#include "keyboard_component.h"
+#include "collider_component.h"
+#include "transform_component.h"
+#include "text_label_component.h"
+#include "projectile_emitter_component.h"
 
 namespace Radix
 {
@@ -157,8 +156,8 @@ namespace Radix
 	{
 		TransformComponent* player_transform = player.GetComponent<TransformComponent>();
 
-		camera.x = player_transform->position.x - (WINDOW_WIDTH / 2);
-		camera.y = player_transform->position.y - (WINDOW_HEIGHT / 2);
+		camera.x = player_transform->position.x - (int)(WINDOW_WIDTH / 2);
+		camera.y = player_transform->position.y - (int)(WINDOW_HEIGHT / 2);
 
 		camera.x = (camera.x < 0) ? 0 : camera.x;
 		camera.y = (camera.y < 0) ? 0 : camera.y;
